refactor(models): move LittlepayConfig to enrollment_littlepay app
this required moving Environment to the common module so as to
not have a circular import.

data migration handles preserving existing LittlepayConfigs over to
the new one.
