+++ benefits/core/admin/transit.py
@@ -5,17 +5,6 @@ from benefits.core import models
 from .users import is_staff_member_or_superuser
 
 
-@admin.register(models.LittlepayConfig)
-class LittlepayConfigAdmin(admin.ModelAdmin):
-    def has_module_permission(self, request):
-        """
-        This controls whether the model shows up in the main list of models.
-        """
-        # we don't want to display LittlepayConfig on the main list.
-        # the user should view it from the TransitAgency.
-        return False
-
-
 @admin.register(models.SwitchioConfig)
 class SwitchioConfigAdmin(admin.ModelAdmin):
     def has_module_permission(self, request):
